% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dt_addDefaultStyleToDataLayer}
\alias{dt_addDefaultStyleToDataLayer}
\title{add a default style (precalculalted and stored in geoserver) for a layer (vector or tiff)}
\usage{
dt_addDefaultStyleToDataLayer(stylename, datalayername, isglobalstyle = FALSE)
}
\arguments{
\item{stylename}{style name (without workspace infomation)}

\item{datalayername}{datalyer name, workspace information will be automatically attached based on devkey account}

\item{isglobalstyle}{if TRUE, workspace information will not be attached. set this to TRUE if the style is hard coded in Geoserver. This is useful for creating common styles (such as shadow intensity, rainfall and temperature) at the application/built-in model level, rather than individual job level}
}
\value{
if success, a published datalayername; otherwise, empty string
}
\description{
add a default style (precalculalted and stored in geoserver) for a layer (vector or tiff)
}
