% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dt_addShp2DataStore}
\alias{dt_addShp2DataStore}
\title{upload a shpfile (zip) to geoserver}
\usage{
dt_addShp2DataStore(filepath)
}
\arguments{
\item{filepath}{}
}
\value{
empty string if success or error message
}
\description{
upload a shpfile (zip) to geoserver
}
