% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dt_createFeatureType}
\alias{dt_createFeatureType}
\title{create a featuretype in geoserver datastore, which makes the uploaded shpfile 'published'}
\usage{
dt_createFeatureType(filename)
}
\arguments{
\item{filename}{}
}
\value{
empty string if success or error message
}
\description{
create a featuretype in geoserver datastore, which makes the uploaded shpfile 'published'
}
