% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dt_createWMSStyle}
\alias{dt_createWMSStyle}
\title{create wms style parameters for a given wfs datalayer}
\usage{
dt_createWMSStyle(
  wfsurl,
  styleprefix = "my_stl_",
  attrname,
  palettename = "Reds",
  colorreverseorder = FALSE,
  colornum = 5,
  classifier = "Jenks",
  bordercolor = "black",
  borderwidth = 1,
  bordervisible = TRUE,
  styletype = "single"
)
}
\arguments{
\item{wfsurl}{wfs url for the data layer}

\item{styleprefix}{a prefix for style name, will make the generated styles in GeoServer more identifiable}

\item{attrname}{the attribute name that the style will be created for}

\item{palettename}{check colors defined under each palette name at http://colorbrewer2.org
acceptable color PaletteNames are: YlOrRd: 9 colors PRGn: 11 colors
PuOr:11 colors RdGy: 11 colors Spectral: 11 colors Grays: 9 colors PuBuGn: 9 colors RdPu: 9 colors BuPu: 9 colors YlOrBr: 9 colors Greens: 9 colors
BuGn: 9 colors Accents: 8 colors GnBu: 9 colors PuRd: 9 colors Purples: 9 colors RdYlGn: 11 colors Paired: 12 colors Blues: 9 colors RdBu: 11 colors
Oranges: 9 colors RdYlBu: 11 colors PuBu: 9 colors OrRd: 9 colors Set3: 12 colors Set2: 8 colors Set1: 9 colors Reds: 9 colors PiYG: 11 colors
Dark2: 8 colors YlGn: 9 colors BrBG: 11 colors YlGnBu: 9 colors
Pastel2: 8 colors Pastel1: 9 colors}

\item{colorreverseorder}{true or false(default), whether to reverse the color order defined in a palette}

\item{colornum}{the number of colors to be classified. default 5.  in range of 2-15}

\item{classifier}{acceptable classifier function names are: Jenks(default), EqualInterval, Quantile, StandardDeviation}

\item{bordercolor}{border color for polygon layers, default "black", can be white, gray or black.}

\item{borderwidth}{border stroke width (integer value) for polygon layers. in range \link{1, 10}, default 1}

\item{bordervisible}{border visibility for polygon layers. TRUE/FALSE, default, TRUE}

\item{styletype}{is styling typle, default "single"}
}
\value{

}
\description{
create wms style parameters for a given wfs datalayer
}
