% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dt_createWorkspace}
\alias{dt_createWorkspace}
\title{create a new workspace}
\usage{
dt_createWorkspace(wsname)
}
\arguments{
\item{wsname}{workspace name}
}
\value{
empty string if success or error message
}
\description{
create a new workspace
}
