% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dt_debugprint}
\alias{dt_debugprint}
\title{debug print}
\usage{
dt_debugprint(arg1)
}
\arguments{
\item{arg1}{A string to be printed}
}
\value{
debug print string started with "==="
}
\description{
debug print
}
\examples{
dt_debugprint("all done")
}
