% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dt_deleteTiffDataLayer}
\alias{dt_deleteTiffDataLayer}
\title{delete a coverage layer based on layername}
\usage{
dt_deleteTiffDataLayer(datalayername)
}
\value{
if success, return the deleted datalayername; otherwise, empty string
}
\description{
delete a coverage layer based on layername
}
