% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dt_getFeatureType}
\alias{dt_getFeatureType}
\title{get all featuretypes defined in the given datastore of the given workspace in geoserver (workspace and datastore are defined the globalGSCredentials)}
\usage{
dt_getFeatureType()
}
\value{
xml string if success or empty string
}
\description{
get all featuretypes defined in the given datastore of the given workspace in geoserver (workspace and datastore are defined the globalGSCredentials)
}
