% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dt_initGeoServerCredentials}
\alias{dt_initGeoServerCredentials}
\title{get GeoServer Credentials from server and initialise variables for current devkey}
\usage{
dt_initGeoServerCredentials(dk)
}
\value{

}
\description{
get GeoServer Credentials from server and initialise variables for current devkey
}
