% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dt_loadGeoJSON2SP}
\alias{dt_loadGeoJSON2SP}
\title{load geojson from url into a sp object}
\usage{
dt_loadGeoJSON2SP(url)
}
\arguments{
\item{url}{A geojson url string}
}
\value{
A sp object
}
\description{
load geojson from url into a sp object
}
