% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dt_loadGeoJSON2SPWithAuth}
\alias{dt_loadGeoJSON2SPWithAuth}
\title{load geojson from url (username/password protected) into a sp object}
\usage{
dt_loadGeoJSON2SPWithAuth(url, username, password)
}
\arguments{
\item{url}{A geojson url string}

\item{username}{username required to access url}

\item{password}{password required to access url}
}
\value{
A sp object
}
\description{
load geojson from url (username/password protected) into a sp object
}
