% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dt_long2UTM}
\alias{dt_long2UTM}
\title{find utm zone for given longitude}
\usage{
dt_long2UTM(long)
}
\arguments{
\item{long}{Longitude}
}
\value{
UTM zone code
}
\description{
find utm zone for given longitude
}
