% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dt_project2WGS84}
\alias{dt_project2WGS84}
\title{reproject(transform) a sp object to a WGS84 (EPSG:4326)}
\usage{
dt_project2WGS84(spobj)
}
\arguments{
\item{spobj}{}
}
\value{
a reprojected sp object
}
\description{
reproject(transform) a sp object to a WGS84 (EPSG:4326)
}
