% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dt_publishTiffDataLayer}
\alias{dt_publishTiffDataLayer}
\title{publish a geotiff to geoserver}
\usage{
dt_publishTiffDataLayer(filepath, datalayername = UUIDgenerate(FALSE))
}
\arguments{
\item{datalayername}{}
}
\value{
if success, a published datalayername; otherwise, empty string
}
\description{
publish a geotiff to geoserver
}
