% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtpluginr-package.R
\docType{package}
\name{dtpluginr-package}
\alias{dtpluginr}
\alias{dtpluginr-package}
\title{dtpluginr: A Fork of DigiTwin's R Plugin}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/asiripanich/dtpluginr}
  \item \url{https://asiripanich.github.io/dtpluginr/}
  \item Report bugs at \url{https://github.com/asiripanich/dtpluginr/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
