% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{utils.createFeatureType}
\alias{utils.createFeatureType}
\title{create a featuretype in geoserver datastore, which makes the uploaded shpfile 'published'}
\usage{
utils.createFeatureType(filename)
}
\arguments{
\item{filename}{}
}
\value{
empty string if success or error message
}
\description{
create a featuretype in geoserver datastore, which makes the uploaded shpfile 'published'
}
