% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{utils.debugprint}
\alias{utils.debugprint}
\title{debug print}
\usage{
utils.debugprint(arg1)
}
\arguments{
\item{arg1}{A string to be printed}
}
\value{
debug print string started with "==="
}
\description{
debug print
}
\examples{
utils.debugprint("all done")
}
