% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{utils.deleteTiffDataLayer}
\alias{utils.deleteTiffDataLayer}
\title{delete a coverage layer based on layername}
\usage{
utils.deleteTiffDataLayer(datalayername)
}
\value{
if success, return the deleted datalayername; otherwise, empty string
}
\description{
delete a coverage layer based on layername
}
