% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{utils.deleteTiffDataStore}
\alias{utils.deleteTiffDataStore}
\title{delete a coverage store (related to current mydevkey) and all layers it contains on geoserver (a bit dangerous)}
\usage{
utils.deleteTiffDataStore()
}
\value{
if success tiff store name; otherwise empty string
}
\description{
delete a coverage store (related to current mydevkey) and all layers it contains on geoserver (a bit dangerous)
}
