% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{utils.df2jsonlist}
\alias{utils.df2jsonlist}
\title{convert data.frame to json list structure}
\usage{
utils.df2jsonlist(df)
}
\arguments{
\item{df}{}
}
\value{
list
}
\description{
convert data.frame to json list structure
}
