% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{utils.getBbox}
\alias{utils.getBbox}
\title{get boundingbox of a sp object}
\usage{
utils.getBbox(spobj)
}
\arguments{
\item{spobj}{}
}
\value{
list(minX, minY, maxX, maxY)
}
\description{
get boundingbox of a sp object
}
