% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{utils.getGeomType}
\alias{utils.getGeomType}
\title{get the geometry type of sp object}
\usage{
utils.getGeomType(spobj)
}
\arguments{
\item{spobj}{}
}
\value{
geomtype, one of Polygon, LineString, Point
}
\description{
get the geometry type of sp object
}
