% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{utils.initGeoServerCredentials}
\alias{utils.initGeoServerCredentials}
\title{get GeoServer Credentials from server and initialise variables for current devkey}
\usage{
utils.initGeoServerCredentials(dk)
}
\value{

}
\description{
get GeoServer Credentials from server and initialise variables for current devkey
}
