% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{utils.loadGeoJSON2DF}
\alias{utils.loadGeoJSON2DF}
\title{load geojson from url into a data.frame object. this is particularly useful to handle geojson wfsurl with null geometry}
\usage{
utils.loadGeoJSON2DF(url)
}
\arguments{
\item{url}{A geojson url string}
}
\value{
A data.frame object
}
\description{
load geojson from url into a data.frame object. this is particularly useful to handle geojson wfsurl with null geometry
}
