% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{utils.long2UTM}
\alias{utils.long2UTM}
\title{find utm zone for given longitude}
\usage{
utils.long2UTM(long)
}
\arguments{
\item{long}{Longitude}
}
\value{
UTM zone code
}
\description{
find utm zone for given longitude
}
