% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{utils.project2UTM}
\alias{utils.project2UTM}
\title{reproject(transform) a sp object to a proper utm crs}
\usage{
utils.project2UTM(spobj)
}
\arguments{
\item{spobj}{}
}
\value{
a reprojected sp object
}
\description{
reproject(transform) a sp object to a proper utm crs
}
