% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{utils.publishSP2GeoServer}
\alias{utils.publishSP2GeoServer}
\title{publish a sp object to geoserver using RESTful API via curl}
\usage{
utils.publishSP2GeoServer(spobj, layerprefix = "my_")
}
\arguments{
\item{spobj}{A sp object. \emph{IMPORTANT: Each column of spobj MUST NOT have attribute name exceeding length (int)10, failed to do so may result in NO Style being generated on GeoServer}}

\item{layerprefix}{a prefix for layer name, will make the generated data layers in GeoServer more identifiable}
}
\value{
A wfs url string of successfully published data layer
}
\description{
publish a sp object to geoserver using RESTful API via curl
}
