% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{utils.publishTiffDataLayer}
\alias{utils.publishTiffDataLayer}
\title{publish a geotiff to geoserver}
\usage{
utils.publishTiffDataLayer(filepath, datalayername = UUIDgenerate(FALSE))
}
\arguments{
\item{filepath}{}

\item{datalayername}{}
}
\value{
if success, a published datalayername; otherwise, empty string
}
\description{
publish a geotiff to geoserver
}
