% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{utils.updateJob}
\alias{utils.updateJob}
\title{sync and update job outputs}
\usage{
utils.updateJob(joboutputs, jobsuccess, jobuuid)
}
\arguments{
\item{joboutputs}{}

\item{jobsuccess}{true or false}

\item{jobuuid}{}
}
\value{

}
\description{
sync and update job outputs
}
